Recreates Database
