+++ db/migrate/20210305020242_create_users.rb
@@ -1,7 +1,6 @@
 class CreateUsers < ActiveRecord::Migration[6.1]
   def change
     create_table :users do |t|
-
       t.string :first_name
       t.string :last_name
       t.string :username
