+++ db/migrate/20210305020340_create_courses.rb
