+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_02_27_132710) do
+ActiveRecord::Schema.define(version: 2021_03_05_020348) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -27,14 +27,6 @@ ActiveRecord::Schema.define(version: 2021_02_27_132710) do
     t.datetime "updated_at", precision: 6, null: false
   end
 
-  create_table "favourites", force: :cascade do |t|
-    t.integer "user_id"
-    t.integer "course_id"
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["user_id", "course_id"], name: "favourites_index", unique: true
-  end
-
   create_table "users", force: :cascade do |t|
     t.string "first_name"
     t.string "last_name"
@@ -47,6 +39,4 @@ ActiveRecord::Schema.define(version: 2021_02_27_132710) do
   end
 
   add_foreign_key "courses", "users", column: "course_author"
-  add_foreign_key "favourites", "courses"
-  add_foreign_key "favourites", "users"
 end
