+++ test/models/course_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require "test_helper"
 
 class CourseTest < ActiveSupport::TestCase
   # test "the truth" do
