+++ test/models/favourite_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require "test_helper"
 
 class FavouriteTest < ActiveSupport::TestCase
   # test "the truth" do
